consumer/requirements.txt
influxdb-client~=1.36.1
platformshconfig~=2.4.0
kafka-python~=2.0.2